ai-api/pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "eventai-api"
version = "1.0.0"
description = "EventAI - Modular AI API for Blur Detection, Face Recognition, and Bib Number OCR"
requires-python = ">=3.11,<3.13"
license = {text = "Proprietary"}

dependencies = [
    # Web framework
    "fastapi>=0.115.0,<1",
    "uvicorn[standard]>=0.34.0,<1",
    "pydantic>=2.10.0,<3",
    "pydantic-settings>=2.7.0,<3",
    "python-multipart>=0.0.18",

    # Image processing
    "opencv-python-headless>=4.10.0,<5",
    "numpy>=2.1.0,<3",
    "Pillow>=11.0.0,<12",

    # ML models
    "insightface>=0.7.3,<0.8",
    "onnxruntime>=1.20.0,<2",
    "paddleocr>=2.9.0,<3",
    "paddlepaddle>=2.6.0,<3",
    "ultralytics>=8.3.0,<9",

    # Database
    "sqlalchemy[asyncio]>=2.0.36,<3",
    "asyncpg>=0.30.0,<1",
    "alembic>=1.14.0,<2",
    "pgvector>=0.3.6,<0.4",

    # Task queue & cache
    "celery[redis]>=5.4.0,<6",
    "redis>=5.2.0,<6",

    # Auth
    "python-jose[cryptography]>=3.3.0,<4",
    "passlib[bcrypt]>=1.7.4,<2",

    # Observability
    "structlog>=24.4.0,<25",
    "prometheus-client>=0.21.0,<1",
    "prometheus-fastapi-instrumentator>=7.0.0,<8",

    # Utilities
    "httpx>=0.28.0,<1",
    "tenacity>=9.0.0,<10",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.20.0,<2",
    "paddlepaddle-gpu>=2.6.0,<3",
]
dev = [
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.25.0,<1",
    "pytest-cov>=6.0.0,<7",
    "httpx>=0.28.0",
    "ruff>=0.9.0,<1",
    "mypy>=1.14.0,<2",
    "pre-commit>=4.0.0,<5",
]
cpp = [
    "pybind11>=2.13.0,<3",
    "scikit-build-core>=0.10.0,<1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
